chore: move get_print to print_utils.py
